move shell component up to lib for usage in both places
