+++ app/src/lib/shell.ts
