+++ app/src/main-process/main.ts
@@ -2,8 +2,8 @@ import { app, Menu, MenuItem, ipcMain, BrowserWindow, autoUpdater, dialog } from
 
 import { AppWindow } from './app-window'
 import { CrashWindow } from './crash-window'
-import { shellNeedsPatching, getEnvironmentFromShell } from './shell'
 import { buildDefaultMenu, MenuEvent, findMenuItemByID, setCrashMenu } from './menu'
+import { shellNeedsPatching, getEnvironmentFromShell } from '../lib/shell'
 import { parseURL } from '../lib/parse-url'
 import { handleSquirrelEvent } from './squirrel-updater'
 import { SharedProcess } from '../shared-process/shared-process'
